Safari-611.1.10/LayoutTests/fast/dom/document-all-undefined.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<script>
description("Tests that document.all masquerades as undefined.");

if (document.all)
    testFailed("ToBoolean() should return false for document.all");
else
    testPassed("ToBoolean() returned false for document.all");

shouldBeTrue("document.all == undefined");
shouldBeFalse("document.all === undefined");
shouldBeTrue("document.all == null");
shouldBeFalse("document.all === null");
shouldBeEqualToString("typeof document.all", "undefined");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
