Safari-609.1.11.1/LayoutTests/fast/dom/document-all-undefined-expected.txt

Tests that document.all masquerades as undefined.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ToBoolean() returned false for document.all
PASS document.all == undefined is true
PASS document.all === undefined is false
PASS document.all == null is true
PASS document.all === null is false
PASS typeof document.all is "undefined"
PASS successfullyParsed is true

TEST COMPLETE
